extracted version compare into VersionComparator with compare method
